+++ pybamm/models/base_model.py
@@ -9,16 +9,6 @@ import warnings
 from collections import OrderedDict
 
 
-class ParamClass:
-    """Class for converting a module of parameters into a class. For pickling."""
-
-    def __init__(self, methods):
-        for k, v in methods.__dict__.items():
-            # don't save module attributes (e.g. pybamm, numpy)
-            if not (k.startswith("__") or inspect.ismodule(v)):
-                self.__dict__[k] = v
-
-
 class BaseModel(object):
     """Base model class for other models to extend.
 
@@ -263,12 +253,7 @@ class BaseModel(object):
 
     @param.setter
     def param(self, values):
-        if values is None:
-            self._param = None
-        else:
-            # convert module into a class
-            # (StackOverflow: https://tinyurl.com/yk3euon3)
-            self._param = ParamClass(values)
+        self._param = values
 
     @property
     def options(self):